Distinguish clipboard from primary selection on X11 systems.
On X11 systems, there are actually three distinct "clipboards", or X
selections, namely the `PRIMARY`, `SECONDARY` and `CLIPBOARD`
selections.  To emulate Vim's behaviour under X11 (see vim documentation
`|quote+|`), we now use the `+` register for accessing the
`CLIPBOARD` selection, and the `*` register for the `PRIMARY` selection.

On non-X11 systems, the distinction is probably moot.  Therefore, we try
to do nothing new, and both the `+` and the `*` registers will access
the same clipboard.